Caching and compression demonstration
By default, the server does not compress any response. We can easily
switch that on via the `spring.compression` namespace and only compress
JSON, css and Javascript.

We can see via Chrome Devtools that the json response is significantly
smaller now. But we could also cache the whole document in the browser
itself. For that we can create an etag that defines the from and to
versions. Because our library is resolving the actual version when it
needs to provide the diff for a SNAPSHOT, the etag is truely unique.