Switch to separate versions provider app
Create a separate `spring-boot-4tw-versions-provider` application that
provides the same feature as our extra controller providing the Spring
Boot versions.

This application can be deployed on CloudFoundry and the Angular app
updated to showcase how it can connect to another service. Note that our
extra app enable Spring 4.2's CORS support so that the browser accepts
a connection to a separate service.